# Faster is_sparse_vector()