--- a/R/type-predicates.R
+++ b/R/type-predicates.R
@@ -34,21 +34,7 @@ NULL
 #' @rdname type-predicates
 #' @export
 is_sparse_vector <- function(x) {
-  res <- .Call(ffi_extract_altrep_class, x)
-  if (is.null(res)) {
-    return(FALSE)
-  }
-  
-  res <- as.character(res[[1]])
-
-  valid <- c(
-    "altrep_sparse_double", 
-    "altrep_sparse_integer", 
-    "altrep_sparse_string",
-    "altrep_sparse_logical"
-  )
-
-  res %in% valid
+  .Call(ffi_is_sparse_vector, x)
 }
 
 #' @rdname type-predicates