--- a/src/init.c
+++ b/src/init.c
@@ -35,6 +35,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"ffi_altrep_sparse_values", (DL_FUNC) &ffi_altrep_sparse_values, 1},
     {"ffi_altrep_sparse_default", (DL_FUNC) &ffi_altrep_sparse_default, 1},
     {"ffi_extract_altrep_class", (DL_FUNC) &ffi_extract_altrep_class, 1},
+    {"ffi_is_sparse_vector", (DL_FUNC) &ffi_is_sparse_vector, 1},
     {NULL, NULL, 0}};
 
 void R_init_sparsevctrs(DllInfo* dll) {