--- a/src/sparse-utils.h
+++ b/src/sparse-utils.h
@@ -25,6 +25,8 @@ bool is_altrep(SEXP x);
 
 SEXP ffi_extract_altrep_class(SEXP x);
 
+SEXP ffi_is_sparse_vector(SEXP x);
+
 R_xlen_t binary_search(int needle, const int* v_haystack, R_xlen_t size);
 
 bool is_index_handleable(SEXP x);