Now all the calculations are done in C.

- if `x` isn't an altrep vector return `FALSE` right away.
- if the package attribute of the altrep. vector is {sparsevctrs} return TRUE.

## With this PR

``` r
library(sparsevctrs)

vec <- sparse_double(1, 1, 1)

bench::mark(
  is_sparse_vector(vec)
)
#> # A tibble: 1 × 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 is_sparse_vector(vec)    164ns    246ns  3177418.    2.26KB        0

bench::mark(
  is_sparse_vector(letters)
)
#> # A tibble: 1 × 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 is_sparse_vector(letters)    123ns    205ns  4795355.    1.92KB        0
```

## Before this PR

``` r
library(sparsevctrs)

vec <- sparse_double(1, 1, 1)

bench::mark(
  is_sparse_vector(vec)
)
#> # A tibble: 1 × 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 is_sparse_vector(vec)    656ns    779ns  1025601.    7.34KB        0

bench::mark(
  is_sparse_vector(letters)
)
#> # A tibble: 1 × 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 is_sparse_vector(letters)    205ns    246ns  3777087.    1.92KB        0
```